+++ cardinal_pythonlib/sql/validation.py
@@ -33,13 +33,177 @@ from typing import Optional, Tuple
 
 
 # =============================================================================
-# SQL types and validation
+# SQL types and validation: constants
 # =============================================================================
 
 # REGEX_INVALID_TABLE_FIELD_CHARS = re.compile("[^a-zA-Z0-9_ ]")
 REGEX_INVALID_TABLE_FIELD_CHARS = re.compile("[^\x20-\x7E]")
 # ... SQL Server is very liberal!
 
+
+# - ANSI: http://jakewheat.github.io/sql-overview/sql-2011-foundation-grammar.html#predefined-type  # noqa
+# - SQL Server: https://support.microsoft.com/en-us/office/equivalent-ansi-sql-data-types-7a0a6bef-ef25-45f9-8a9a-3c5f21b5c65d  # noqa
+#   - Note that ANSI "BIT" is SQL Server "BINARY".
+# - MySQL: https://dev.mysql.com/doc/refman/8.0/en/data-types.html
+# - PostgreSQL: https://www.postgresql.org/docs/9.5/datatype.html
+
+SQLTYPES_INTEGER = (
+    "BIGINT",  # ANSI
+    "BIGSERIAL",  # PostgreSQL
+    "BYTE",  # SQL Server
+    "COUNTER",  # SQL Server
+    "INT",  # ANSI; synonym for INTEGER
+    "INT2",  # PostgreSQL
+    "INT4",  # PostgreSQL synonym for INT
+    "INTEGER",  # ANSI
+    "INTEGER1",  # SQL Server
+    "INTEGER2",  # SQL Server
+    "INTEGER4",  # SQL Server
+    "LONG",  # SQL Server
+    "MEDIUMINT",  # MySQL
+    "SERIAL",  # PostgreSQL
+    "SERIAL2",  # PostgreSQL
+    "SERIAL4",  # PostgreSQL
+    "SHORT",  # SQL Server
+    "SMALLINT",  # ANSI
+    "SMALLSERIAL",  # PostgreSQL
+    "TINYINT",  # SQL Server, MySQL
+)
+SQLTYPES_FLOAT = (
+    "DOUBLE PRECISION",  # ANSI (8 bytes)
+    "DOUBLE",  # SQL Server, MySQL; synonym for DOUBLE PRECISION
+    "FLOAT",  # ANSI
+    "FLOAT4",  # SQL Server
+    "FLOAT8",  # SQL Server
+    "IEEEDOUBLE",  # SQL Server
+    "IEEESINGLE",  # SQL Server
+    "NUMBER",  # SQL Server
+    "REAL",  # ANSI (though MySQL says it is not standard)
+    "SINGLE",  # SQL Server
+)
+SQLTYPES_OTHER_NUMERIC = (
+    "BIT VARYING",  # ANSI
+    "BIT",  # ANSI
+    "BOOL",  # MySQL synonym for BOOLEAN or TINYINT(1)
+    "BOOLEAN",  # ANSI
+    "DEC",  # ANSI; synonym for DECIMAL
+    "DECIMAL",  # ANSI
+    "FIXED",  # MySQL; synonym for DECIMAL
+    "LOGICAL",  # SQL Server
+    "LOGICAL1",  # SQL Server
+    "NUMERIC",  # ANSI; synonym for DECIMAL
+    "VARBIT",  # PostgreSQL synonym for BIT VARYING
+    "YESNO",  # SQL Server
+)
+SQLTYPES_TEXT = (
+    "ALPHANUMERIC",  # SQL Server
+    "CHAR LARGE OBJECT",  # ANSI
+    "CHAR VARYING",  # ANSI
+    "CHAR",  # ANSI
+    "CHARACTER LARGE OBJECT",  # ANSI
+    "CHARACTER VARYING",  # ANSI
+    "CHARACTER",  # ANSI
+    "CLOB",  # ANSI
+    "ENUM",  # MySQL
+    "LONGCHAR",  # SQL Server
+    "LONGTEXT",  # SQL Server, MySQL
+    "MEDIUMTEXT",  # MySQL
+    "MEMO",  # SQL Server
+    "NATIONAL CHAR VARYING",  # ANSI
+    "NATIONAL CHAR",  # ANSI
+    "NATIONAL CHARACTER LARGE OBJECT",  # ANSI
+    "NATIONAL CHARACTER VARYING",  # ANSI
+    "NATIONAL CHARACTER",  # ANSI
+    "NCHAR LARGE OBJECT",  # ANSI
+    "NCHAR VARYING",  # ANSI
+    "NCHAR",  # ANSI
+    "NCLOB",  # ANSI
+    "NOTE",  # SQL Server
+    "NTEXT",  # SQL Server
+    "NVARCHAR",  # SQL Server
+    "SET",  # MySQL
+    "STRING",  # SQL Server
+    "TEXT",  # SQL Server, MySQL
+    "TINYTEXT",  # MySQL
+    "VARCHAR",  # ANSI
+)
+SQLTYPES_BINARY = (
+    "BINARY LARGE OBJECT",  # ANSI
+    "BINARY VARYING",  # ANSI
+    "BINARY",  # ANSI
+    "BLOB",  # ANSI
+    "BYTEA",  # PostgreSQL ("byte array")
+    "GENERAL",  # SQL Server
+    "IMAGE",  # SQL Server
+    "LONGBINARY",  # SQL Server
+    "LONGBLOB",  # MySQL
+    "MEDIUMBLOB",  # MySQL
+    "OLEOBJECT",  # SQL Server
+    "TINYBLOB",  # MySQL
+    "VARBINARY",  # ANSI
+)
+SQLTYPES_WITH_DATE = (
+    "DATE",  # ANSI
+    "DATETIME",  # SQL Server, MySQL
+    "TIMESTAMP",  # ANSI
+)
+SQLTYPES_DATETIME_OTHER = (
+    "INTERVAL",  # ANSI (not always supported); PostgreSQL
+    "TIME",  # ANSI
+    "YEAR",  # MySQL
+)
+SQLTYPES_OTHER = (
+    "BOX",  # PostgreSQL
+    "CIDR",  # PostgreSQL
+    "CIRCLE",  # PostgreSQL
+    "CURRENCY",  # SQL Server
+    "GUID",  # SQL Server
+    "INET",  # PostgreSQL
+    "JSON",  # MySQL, PostgreSQL
+    "JSONB",  # PostgreSQL
+    "LINE",  # PostgreSQL
+    "LSEG",  # PostgreSQL
+    "MACADDR",  # PostgreSQL
+    "MONEY",  # SQL Server, PostgreSQL
+    "PATH",  # PostgreSQL
+    "PG_LSN",  # PostgreSQL
+    "POINT",  # PostgreSQL
+    "POLYGON",  # PostgreSQL
+    "TSQUERY",  # PostgreSQL
+    "TSVECTOR",  # PostgreSQL
+    "TXID_SNAPSHOT",  # PostgreSQL
+    "UNIQUEIDENTIFIER",  # SQL Server
+    "UUID",  # PostgreSQL
+    "XML",  # PostgreSQL
+)
+
+SQLTYPES_DATETIME_ALL = SQLTYPES_WITH_DATE + SQLTYPES_DATETIME_OTHER
+SQLTYPES_ALL = (
+    SQLTYPES_INTEGER +
+    SQLTYPES_FLOAT +
+    SQLTYPES_OTHER_NUMERIC +
+    SQLTYPES_TEXT +
+    SQLTYPES_BINARY +
+    SQLTYPES_DATETIME_ALL +
+    SQLTYPES_OTHER
+)
+SQLTYPES_NOT_TEXT = (
+    SQLTYPES_INTEGER +
+    SQLTYPES_FLOAT +
+    SQLTYPES_OTHER_NUMERIC +
+    SQLTYPES_DATETIME_ALL +
+    SQLTYPES_OTHER
+)
+SQLTYPES_NUMERIC = (
+    SQLTYPES_INTEGER +
+    SQLTYPES_FLOAT +
+    SQLTYPES_OTHER_NUMERIC
+)
+
+
+# =============================================================================
+# SQL types and validation: functions
+# =============================================================================
 
 def is_valid_field_name(f: Optional[str]) -> bool:
     if not f:
@@ -61,55 +225,6 @@ def ensure_valid_field_name(f: Optional[str]) -> None:
 def ensure_valid_table_name(t: Optional[str]) -> None:
     if not is_valid_table_name(t):
         raise ValueError(f"Table name invalid: {t}")
-
-
-SQLTYPES_INTEGER = [
-    "INT", "INTEGER",
-    "TINYINT", "SMALLINT", "MEDIUMINT", "BIGINT",
-]
-SQLTYPES_FLOAT = [
-    "DOUBLE", "FLOAT",
-]
-SQLTYPES_OTHER_NUMERIC = [
-    "BIT", "BOOL", "BOOLEAN", "DEC", "DECIMAL",
-]
-SQLTYPES_TEXT = [
-    "CHAR", "VARCHAR", "NVARCHAR",
-    "TINYTEXT", "TEXT", "NTEXT", "MEDIUMTEXT", "LONGTEXT",
-]
-SQLTYPES_BINARY = [
-    "BINARY", "BLOB", "IMAGE", "LONGBLOB", "VARBINARY",
-]
-
-SQLTYPES_WITH_DATE = [
-    "DATE", "DATETIME", "TIME", "TIMESTAMP",
-]
-SQLTYPES_DATETIME_OTHER = [
-    "TIME", "YEAR",
-]
-SQLTYPES_DATETIME_ALL = SQLTYPES_WITH_DATE + SQLTYPES_DATETIME_OTHER
-
-SQLTYPES_ALL = (
-    SQLTYPES_INTEGER +
-    SQLTYPES_FLOAT +
-    SQLTYPES_OTHER_NUMERIC +
-    SQLTYPES_TEXT +
-    SQLTYPES_BINARY +
-    SQLTYPES_DATETIME_ALL
-)
-# Could be more comprehensive!
-
-SQLTYPES_NOT_TEXT = (
-    SQLTYPES_INTEGER +
-    SQLTYPES_FLOAT +
-    SQLTYPES_OTHER_NUMERIC +
-    SQLTYPES_DATETIME_ALL
-)
-SQLTYPES_NUMERIC = (
-    SQLTYPES_INTEGER +
-    SQLTYPES_FLOAT +
-    SQLTYPES_OTHER_NUMERIC
-)
 
 
 def split_long_sqltype(datatype_long: str) -> Tuple[str, Optional[int]]: